+++ doc/changelog.rst
@@ -26,6 +26,7 @@ Enhancements
 * Add :meth:`mne_nirs.channels.drop_detectors`. By `Robert Luke`_.
 * Add :meth:`mne_nirs.channels.pick_sources`. By `Robert Luke`_.
 * Add :meth:`mne_nirs.channels.pick_detectors`. By `Robert Luke`_.
+* Add :meth:`mne_nirs.preprocessing.quanitfy_mayer_fooof`. By `Robert Luke`_.
 
 
 v0.1.1